maskan_api/requirements.txt
django>=5.0,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
django-ratelimit>=4.1
python-decouple>=3.8
psycopg2-binary>=2.9
Pillow>=10.0
